Normalizes node features so they use LDK NodeFeatures
Currently we were using a pretty hacky way of parsing LND features that only
parsed the keysend feature (given is the only we care about atm).

This PR properly loads the feature vector returned by LND's GetNodeInfo and
parses it so a complete NodeFeatures object can be constructed.

Also sets NodeInfo::features to NodeFeatures for consistency

This is based on LDK's code for parsing features, there's currently an open
PR to merge this upstream, so we can simplify if that gets merged.