sim-all: Allows aliases within node definition in the config file
Crates a new enum, `NodeId`, which is used to parse the node identifier from
the config file to either a `PublicKey` or an `Alias`. Internally, we will identify
nodes as `PublicKey`s, so we add do further checks when creating {Cln, Lnd}Connection.

Moreover,  we make sure that the provided variant matches the one from the backend.
In case of a `PublicKey`, if the assertion fails the code will return a `ValidationError`.
In the case of an `Alias`, it will simply log a warning.

The rationale for not failing if the provided alias does not match the one from
the backend is that we want to allow user to identify nodes using whatever name
they rather (alice/bob/carol/..., n1, n2, n3, ...), so there is not strong requirement
for aliases to match (but it feels right to let the user know, just in case).

Finally, if an identifier is used more than once, the simulation will also fail with a
`ValidationError`, independently of the identifier variant.