sim-lib: separate consumer and producer spin up into dedicated methods
When we add random activity generation, we're going to want consumers
for every node that we have execution on. Previously, we only ran
consumers for nodes that are listed as the source in activity
description. This change separates consumer and producer spin up and
allows the caller to specify which nodes to run consumers for.