sim-cli: simplifies main.rs by deduplicating the node creation
Node creation logic was mostly duplicated, creates a node based on LightningNode
instead of on its implementations to reduce the code replication.

I'd be nice the be able to do this with a different container than Arc<Mutex<T>>,
since that will save us from having to acquire the mutex after creating `node`.
However, I don't know of any other container that is is send and which size is
known at compile time. We may need to ask this to some elder rust wizards