sim-lib: Adds LightningNode::get_network
The current approach to get the network from a node stores the information in `NodeInfo`.

However, this is troublesome going forward given we will start using `NodeInfo`
to store information about nodes we do not control (and from which we cannot
query the network). Furthermore, the network information is worthless once we've
sanity-checked that we are running on a network we support and all nodes are running
on the same one, so there is no point keeping that around