sim-all: renames config file to sim file
Currently we are using the term config file for the file that contains the
simulation details. This is kind of fine for now, but it will get messy down the
line when we want to start storing default values for the CLI parameters in a config
file. Furthermore, we are requiring the config file to be passed, but `--config` is
required.

This both renames config to sim, and makes the parameter positional instead of named.