sim-node: add channel state tracking for channels
This commit adds a ChannelState struct which is used to track the
policy and state of a channel in the *outgoing* direction. This will
be used to check forwards against the node's advertised policy and
track the movement of outgoing HTLCs through the channel.

Note that we choose to implement this state *unidirectionally*, so a
single channel will be represented by two ChannelState structs (one in
each direction).