sim-lib: only use start delay if explicitly specified
As is, we'll run with a 0 second start delay for:
- Random activities
- Defined activities with no `start_secs` specified

This very likely isn't the intent for defined activities (people should
set `start_secs=0` if they want their payments to fire immediately),
and doesn't work for random activity because we just fire all payments
on start.

Instead, we switch over to using our payment wait as the start delay in
the absence of this field being set for defined activities and always
for random activities.