multi: add cfg struct to simulation
Nodes and activity are not included here, because we want to be
able to create our config struct separately from parsing of nodes
and activity. This reasoning will become more apparent in future
commits when we add two different running modes for the simulator
(regular operation and simulated network mode).