docker: fix base image to 1.85
Changes to later versions of rust have broken our dockerfile. This
change fixes our base image to the version of rust we're currently
using, where the break is not present.

We'll need to fix this in future when we bump versions, but for now
we make a minimum change to be able to cut the 2.4 release.