simln-lib/test: speed up random payment deterministic test clock
Speed up this test by using our sim clock that can speed up the
simulation. The downside of this clock is that we may stop one payment
over/under if we try to match exactly to the number of payments that
we expect in a given period of time.

Instead, we generate our set of expected payment and then run the
simulation for much longer than we need to get this expected list, so
we should always get at least this subset. We then just assert that the
payments we get exactly match this first set.